unify MessageCell
