+++ feature/Messages/Sources/Messages/Views/ConversationView.swift
@@ -114,70 +114,6 @@ private struct ConversationDaySection: View {
     }
 }
 
-private struct MessageCell: View {
-
-    @EnvironmentObject var navigationController: NavigationController
-
-    @ObservedObject var viewModel: ConversationViewModel
-
-    @State private var showMessageActionSheet = false
-    @State private var isPressed = false
-
-    let message: Message
-    let conversationPath: ConversationPath
-    let showHeader: Bool
-
-    var body: some View {
-        HStack(alignment: .top, spacing: .l) {
-            Image(systemName: "person")
-                .resizable()
-                .scaledToFit()
-                .frame(width: 30, height: 30)
-                .padding(.top, .s)
-                .opacity(showHeader ? 1 : 0)
-            VStack(alignment: .leading, spacing: .m) {
-                if showHeader {
-                    HStack(alignment: .bottom, spacing: .m) {
-                        Text(message.author?.name ?? "")
-                            .bold()
-                        if let creationDate = message.creationDate {
-                            Text(creationDate, formatter: DateFormatter.timeOnly)
-                                .font(.caption)
-                        }
-                    }
-                }
-                ArtemisMarkdownView(string: message.content ?? "")
-                ReactionsView(viewModel: viewModel, message: message, showEmojiAddButton: false)
-                if let answerCount = message.answers?.count,
-                   answerCount > 0 {
-                    Button(R.string.localizable.replyAction(answerCount)) {
-                        navigationController.path.append(MessagePath(message: message, coursePath: conversationPath.coursePath, conversationPath: conversationPath))
-                    }
-                }
-            }.id(message.id)
-            Spacer()
-        }
-            .padding(.horizontal, .l)
-            .contentShape(Rectangle())
-            .background(isPressed ? Color.Artemis.messsageCellPressed : Color.clear)
-            .onTapGesture {
-                print("This somehow fixes scrolling...")
-            }
-            .onLongPressGesture(minimumDuration: 0.1, maximumDistance: 30, perform: {
-                let impactMed = UIImpactFeedbackGenerator(style: .heavy)
-                impactMed.impactOccurred()
-                showMessageActionSheet = true
-                isPressed = false
-            }, onPressingChanged: { pressed in
-                isPressed = pressed
-            })
-            .sheet(isPresented: $showMessageActionSheet) {
-                MessageActionSheet(message: message, conversationPath: conversationPath)
-                    .presentationDetents([.height(350), .large])
-            }
-    }
-}
-
 extension Date: Identifiable {
     public var id: Date {
         return self
